block group falling multiple spaces at once fix
previously, if a block group of size > 3 had n empty tiles below it, it would fall down one tile, then immediately extinguish itself. this change should allow that group to fall until it can fall no longer, and THEN extinguish. this does not address the case where a block group falls down one tile and is above another group that is in the process of extinguishing - in that case the group will still extinguish itself immediately instead of waiting for the below group to finish extinguishing and continue falling.